Fix a timestamp-handling bug
If there's a timestamp in the middle of a vote, we should still ignore it. Fixed by having the vote regex check for a signature on the line where it stops, and by having the signature parser use the last line of the vote (not the whole text) as input. Added a new test case to check this.