Fix vote parsing bug
When the preprocessor is extending the text chunk it's currently working on while it's looking for votes with missing bold parts, sometimes it extends over a section boundary (e.g. while looking for the end of an oppose vote, it extends into the neutral section). I added a new check for full headers inside the vote text while the preprocessor is doing the extension.