Fix silly bug in RfX preprocessing
I missed catching spaces in the header regex, so the RfX preprocessor
wouldn't run if the headers were written as `===== Support =====`
instead of `=====Support=====`. The current revision of {{RfA}} doesn't
have the spaces, but Eostrix's does, so I call shenanigans. But that's
why I haven't noticed it until now, I guess.